+++ samples/Sample_05_Chart_Line.php
@@ -3,12 +3,8 @@
 include_once 'Sample_Header.php';
 
 use PhpOffice\PhpPresentation\PhpPresentation;
-use PhpOffice\PhpPresentation\Shape\Chart\Type\Bar3D;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Line;
-use PhpOffice\PhpPresentation\Shape\Chart\Type\Pie3D;
-use PhpOffice\PhpPresentation\Shape\Chart\Type\Scatter;
 use PhpOffice\PhpPresentation\Shape\Chart\Series;
-use PhpOffice\PhpPresentation\Style\Alignment;
 use PhpOffice\PhpPresentation\Style\Border;
 use PhpOffice\PhpPresentation\Style\Color;
 use PhpOffice\PhpPresentation\Style\Fill;
@@ -125,7 +121,6 @@ $shape2->getPlotArea()->setType($lineChart2);
 $shape2->getPlotArea()->getAxisY()->setFormatCode('#,##0');
 $currentSlide->addShape($shape2);
 
-
 // Create templated slide
 echo EOL . date('H:i:s') . ' Create templated slide' . EOL;
 $currentSlide = createTemplatedSlide($objPHPPresentation);
@@ -152,6 +147,18 @@ $shape3->getPlotArea()->setType($lineChart3);
 $shape3->getPlotArea()->getAxisX()->setMajorGridlines($oGridLines1);
 $shape3->getPlotArea()->getAxisY()->setMinorGridlines($oGridLines2);
 $currentSlide->addShape($shape3);
+
+// Create templated slide
+echo EOL . date('H:i:s') . ' Create templated slide' . EOL;
+$currentSlide = createTemplatedSlide($objPHPPresentation);
+
+// Create a shape (chart)
+echo date('H:i:s') . ' Create a shape (chart3)' . EOL;
+echo date('H:i:s') . ' Feature : Gridlines' . EOL;
+$shape4 = clone $shape;
+$shape4->getPlotArea()->getAxisY()->setMinBounds(5);
+$shape4->getPlotArea()->getAxisY()->setMaxBounds(20);
+$currentSlide->addShape($shape4);
 // Save file
 echo EOL . write($objPHPPresentation, basename(__FILE__, '.php'), $writers);